backend/requirements.txt
Django>=4.2,<5
djangorestframework
psycopg2-binary
gunicorn
aiogram
python-dotenv
requests
httpx>=0.25.0